Move example commands into unit test
